Improve photo taking and selection behaviour
New features:

Taking and selecting photos no longer automatically dismiss the modal and call the completion handler.

Selecting a photo from the library selects the relevant photo in the collection view if it's there, or makes it appear as the first item and selects it.

Taking a photo saves the photo to the camera roll and the subsequent update to the photo library is handled and the collection view is updated.

New photos will optinoally be saved in an album. The name of the album is configurable.

New photos or selected photos from library are automatically selected and previous selection is discarded.

Bug fixes

- Status bar is hidden before opening camera utilising modalPresentationCapturesStatusBarAppearance

Known issues

- Small gap on the left hand side of collection view due to secion 0 being empty and using sectionInset on the collection view layout